[orchestration-engine] - Split out UIKitConsumerTests into logical concerns - TT RAG
